+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.2]
 ## [Unreleased]
+## [1.1.0]
+### Added
+- Admin role flag for users
+- Blazer admin dashboard
+
+## [1.0.2]
 ### Fixed
 - Populate sendgrid credentials
 - Add ImageMagick to provisioning scripts
@@ -18,7 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 Initial release supporting basic user accounts and vehicle information
 
-[Unreleased]: https://github.com/elepedus/krmb/compare/1.0.2...HEAD
+[Unreleased]: https://github.com/elepedus/krmb/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/elepedus/krmb/compare/1.1.0...1.0.2
 [1.0.2]: https://github.com/elepedus/krmb/compare/1.0.2...1.0.1
 [1.0.1]: https://github.com/elepedus/krmb/compare/1.0.1...1.0.0
 [1.0.0]: https://github.com/elepedus/krmb/releases/tag/1.0.0